#147. e2e testing

+++ src/review/review.service.ts
@@ -1,9 +1,10 @@
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Review, ReviewDocument } from './review.model';
 import { Model } from 'mongoose';
-import { CreateBookDto } from 'src/books/dto/create-book.dto';
-import { CreateReviewDto, EditReviewDto } from './review.dto';
+import { CreateReviewDto, EditReviewDto } from './dto/review.dto';
+import { Review, ReviewDocument } from './review.model';
 
+@Injectable()
 export class ReviewService {
   constructor(@InjectModel(Review.name) private reviewModel: Model<ReviewDocument>) {}
 
@@ -14,19 +15,22 @@ export class ReviewService {
   }
 
   async deleteReview(reviewId: string) {
-    await this.reviewModel.findByIdAndRemove(reviewId);
+    const isReview = await this.reviewModel.findById(reviewId);
 
-    return reviewId;
+    if (!isReview) throw new NotFoundException('Review with id not found');
+
+    const review = await this.reviewModel.findByIdAndRemove(reviewId);
+
+    return review._id;
   }
 
   async editReview(reviewId: string, dto: EditReviewDto) {
+    console.log(reviewId);
+
     const review = await this.reviewModel.findByIdAndUpdate(
       reviewId,
       {
-        $set: {
-          rating: dto.rating,
-          summary: dto.summary,
-        },
+        $set: { rating: dto.rating, summary: dto.summary },
       },
       { new: true },
     );
@@ -35,8 +39,8 @@ export class ReviewService {
   }
 
   async getReview(courseId: string) {
-    const reviews = await this.reviewModel.find({ course: courseId });
+    const review = await this.reviewModel.find({ course: courseId });
 
-    return reviews; 
+    return review;
   }
 }
